requirements.txt
absl-py
amqp
ann-visualizer
argon2-cffi
asgiref
astor
astunparse
async-generator
attrs
autopep8
backcall
bcrypt
beautifulsoup4
billiard
bleach
bs4
CacheControl
cachetools
celery
celery-progress
certifi
cffi
chardet
click
click-didyoumean
click-repl
colorama
cycler
decorator
defusedxml
distro
dj-database-url
Django
django-cleanup
django-cors-headers
django-extensions
django-heroku
django-pandas
django-phonenumber-field
django-rest-auth
djangorestframework
djongo
dnspython
entrypoints
et-xmlfile
firebase-admin
flatbuffers
gast
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
graphviz
greenlet
grpcio
gunicorn
h5py
httplib2
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
isort
jdcal
jedi
Jinja2
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
Keras
Keras-Applications
keras-nightly
Keras-Preprocessing
kiwisolver
kombu
lib
Markdown
MarkupSafe
matplotlib
matplotlib-inline
missingno
mistune
mlxtend
msgpack
mysql
mysqlclient
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
numpy
oauthlib
openpyxl
opt-einsum
packaging
pandas
pandasql
pandocfilters
parso
phonenumbers
pickleshare
Pillow
plotly
prometheus-client
prompt-toolkit
proto-plus
protobuf
psycopg2
pyasn1
pyasn1-modules
pycodestyle
pycparser
Pygments
pymongo
PyMySQL
pyodbc
pyparsing
PyPDF2
pyreadstat
pyrsistent
python-bsonjs
python-dateutil
python-snappy
pytz
pywin32
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
redis
regex
requests
requests-oauthlib
retrying
rsa
scikit-learn
scipy
seaborn
Send2Trash
six
soupsieve
SQLAlchemy
sqlparse
squarify
tabula
tabula-py
tenacity
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
terminado
testpath
threadpoolctl
toml
tornado
tqdm
traitlets
typing-extensions
tzlocal
uritemplate
urllib3
vine
wcwidth
webencodings
Werkzeug
whitenoise
widgetsnbextension
wincertstore
winkerberos
wordcloud
wrapt
xlrd
zipp